Add provisioning for aliases

+++ lib/Service/Provisioning/Manager.php
@@ -24,6 +24,8 @@ declare(strict_types=1);
 namespace OCA\Mail\Service\Provisioning;
 
 use Horde_Mail_Rfc822_Address;
+use OCA\Mail\Db\Alias;
+use OCA\Mail\Db\AliasMapper;
 use OCA\Mail\Db\MailAccount;
 use OCA\Mail\Db\MailAccountMapper;
 use OCA\Mail\Db\Provisioning;
@@ -33,6 +35,8 @@ use OCP\AppFramework\Db\DoesNotExistException;
 use OCP\AppFramework\Db\MultipleObjectsReturnedException;
 use OCP\IUser;
 use OCP\IUserManager;
+use OCP\LDAP\ILDAPProvider;
+use OCP\LDAP\ILDAPProviderFactory;
 use OCP\Security\ICrypto;
 use Psr\Log\LoggerInterface;
 
@@ -50,6 +54,12 @@ class Manager {
 	/** @var ICrypto */
 	private $crypto;
 
+	/** @var ILDAPProviderFactory */
+	private $ldapProviderFactory;
+
+	/** @var AliasMapper */
+	private $aliasMapper;
+
 	/** @var LoggerInterface */
 	private $logger;
 
@@ -57,11 +67,15 @@ class Manager {
 								ProvisioningMapper $provisioningMapper,
 								MailAccountMapper $mailAccountMapper,
 								ICrypto $crypto,
+								ILDAPProviderFactory $ldapProviderFactory,
+								AliasMapper $aliasMapper,
 								LoggerInterface $logger) {
 		$this->userManager = $userManager;
 		$this->provisioningMapper = $provisioningMapper;
 		$this->mailAccountMapper = $mailAccountMapper;
 		$this->crypto = $crypto;
+		$this->ldapProviderFactory = $ldapProviderFactory;
+		$this->aliasMapper = $aliasMapper;
 		$this->logger = $logger;
 	}
 
@@ -85,6 +99,75 @@ class Manager {
 	}
 
 	/**
+	 * Delete orphaned aliases for the given account.
+	 *
+	 * A alias is orphaned if not listed in newAliases anymore
+	 * (=> the provisioning configuration does contain it anymore)
+	 *
+	 * Exception for Nextcloud 20: \Doctrine\DBAL\DBALException
+	 * Exception for Nextcloud 21 and newer: \OCP\DB\Exception
+	 *
+	 * @TODO: Change throws to \OCP\DB\Exception once Mail requires Nextcloud 21 or above
+	 *
+	 * @throws \Exception
+	 */
+	private function deleteOrphanedAliases(string $userId, int $accountId, array $newAliases): void {
+		$existingAliases = $this->aliasMapper->findAll($accountId, $userId);
+		foreach ($existingAliases as $existingAlias) {
+			if (!in_array($existingAlias->getAlias(), $newAliases, true)) {
+				$this->aliasMapper->delete($existingAlias);
+			}
+		}
+	}
+
+	/**
+	 * Create new aliases for the given account.
+	 *
+	 * Exception for Nextcloud 20: \Doctrine\DBAL\DBALException
+	 * Exception for Nextcloud 21 and newer: \OCP\DB\Exception
+	 *
+	 * @TODO: Change throws to \OCP\DB\Exception once Mail requires Nextcloud 21 or above
+	 *
+	 * @throws \Exception
+	 */
+	private function createNewAliases(string $userId, int $accountId, array $newAliases, string $displayName): void {
+		foreach ($newAliases as $newAlias) {
+			try {
+				$this->aliasMapper->findByAlias($newAlias, $userId);
+			} catch (DoesNotExistException $e) {
+				$alias = new Alias();
+				$alias->setAccountId($accountId);
+				$alias->setName($displayName);
+				$alias->setAlias($newAlias);
+				$this->aliasMapper->insert($alias);
+			}
+		}
+	}
+
+	/**
+	 * @throws \Exception if user id was not found in LDAP
+	 *
+	 * @TODO: Remove psalm-suppress once Mail requires Nextcloud 22 or above
+	 */
+	public function ldapAliasesIntegration(Provisioning $provisioning, IUser $user): Provisioning {
+		if ($user->getBackendClassName() !== 'LDAP' || $provisioning->getLdapAliasesProvisioning() === false || empty($provisioning->getLdapAliasesAttribute())) {
+			return $provisioning;
+		}
+
+		/** @psalm-suppress UndefinedInterfaceMethod */
+		if ($this->ldapProviderFactory->isAvailable() === false) {
+			$this->logger->debug('Request to provision mail aliases but ldap not available');
+			return $provisioning;
+		}
+
+		$ldapProvider = $this->ldapProviderFactory->getLDAPProvider();
+		/** @psalm-suppress UndefinedInterfaceMethod */
+		$provisioning->setAliases($ldapProvider->getMultiValueUserAttribute($user->getUID(), $provisioning->getLdapAliasesAttribute()));
+
+		return $provisioning;
+	}
+
+	/**
 	 * @param Provisioning[] $provisionings
 	 */
 	public function provisionSingleUser(array $provisionings, IUser $user): bool {
@@ -96,57 +179,69 @@ class Manager {
 
 		try {
 			// TODO: match by UID only, catch multiple objects returned below and delete all those accounts
-			$existing = $this->mailAccountMapper->findProvisionedAccount($user);
+			$mailAccount = $this->mailAccountMapper->findProvisionedAccount($user);
 
-			$this->mailAccountMapper->update(
-				$this->updateAccount($user, $existing, $provisioning)
+			$mailAccount = $this->mailAccountMapper->update(
+				$this->updateAccount($user, $mailAccount, $provisioning)
 			);
-			return true;
-		} catch (DoesNotExistException $e) {
+		} catch (DoesNotExistException | MultipleObjectsReturnedException $e) {
+			if ($e instanceof MultipleObjectsReturnedException) {
+				// This is unlikely to happen but not impossible.
+				// Let's wipe any existing accounts and start fresh
+				$this->aliasMapper->deleteProvisionedAliasesByUid($user->getUID());
+				$this->mailAccountMapper->deleteProvisionedAccountsByUid($user->getUID());
+			}
+
 			// Fine, then we create a new one
-			$new = new MailAccount();
-			$new->setUserId($user->getUID());
+			$mailAccount = new MailAccount();
+			$mailAccount->setUserId($user->getUID());
 
-			$this->mailAccountMapper->insert(
-				$this->updateAccount($user, $new, $provisioning)
+			$mailAccount = $this->mailAccountMapper->insert(
+				$this->updateAccount($user, $mailAccount, $provisioning)
 			);
-			return true;
-		} catch (MultipleObjectsReturnedException $e) {
-			// This is unlikely to happen but not impossible.
-			// Let's wipe any existing accounts and start fresh
-			$this->mailAccountMapper->deleteProvisionedAccountsByUid($user->getUID());
-
-			$new = new MailAccount();
-			$new->setUserId($user->getUID());
-
-			$this->mailAccountMapper->insert(
-				$this->updateAccount($user, $new, $provisioning)
-			);
-			return true;
 		}
-		return false;
+
+		// @TODO: Remove method_exists once Mail requires Nextcloud 22 or above
+		if (method_exists(ILDAPProvider::class, 'getMultiValueUserAttribute')) {
+			try {
+				$provisioning = $this->ldapAliasesIntegration($provisioning, $user);
+			} catch (\Throwable $e) {
+				$this->logger->warning('Request to provision mail aliases failed', ['exception' => $e]);
+				// return here to avoid provisioning of aliases.
+				return true;
+			}
+
+			try {
+				$this->deleteOrphanedAliases($user->getUID(), $mailAccount->getId(), $provisioning->getAliases());
+			} catch (\Throwable $e) {
+				$this->logger->warning('Deleting orphaned aliases failed', ['exception' => $e]);
+			}
+
+			try {
+				$this->createNewAliases($user->getUID(), $mailAccount->getId(), $provisioning->getAliases(), $user->getDisplayName());
+			} catch (\Throwable $e) {
+				$this->logger->warning('Creating new aliases failed', ['exception' => $e]);
+			}
+		}
+
+		return true;
 	}
 
+	/**
+	 * @throws ValidationException
+	 * @throws \Exception
+	 */
 	public function newProvisioning(array $data): void {
-		try {
-			$provisioning = $this->provisioningMapper->validate(
-				$data
-			);
-		} catch (ValidationException $e) {
-			throw $e;
-		}
+		$provisioning = $this->provisioningMapper->validate($data);
 		$this->provisioningMapper->insert($provisioning);
 	}
 
+	/**
+	 * @throws ValidationException
+	 * @throws \Exception
+	 */
 	public function updateProvisioning(array $data): void {
-		try {
-			$provisioning = $this->provisioningMapper->validate(
-				$data
-			);
-		} catch (ValidationException $e) {
-			throw $e;
-		}
-
+		$provisioning = $this->provisioningMapper->validate($data);
 		$this->provisioningMapper->update($provisioning);
 	}
